+++ src/main/java/roomescape/controller/ReservationController.java
@@ -2,29 +2,25 @@ package roomescape.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import roomescape.dao.ReservationTimeDao;
 import roomescape.domain.Reservation;
-import roomescape.dao.ReservationDao;
-import roomescape.domain.ReservationTime;
 import roomescape.dto.ReservationDetailDto;
 import roomescape.dto.ReservationSaveDto;
+import roomescape.service.ReservationService;
 
 import java.util.List;
 
 @RestController
 public class ReservationController {
 
-    private final ReservationDao reservationDao;
-    private final ReservationTimeDao reservationTimeDao;
+    private final ReservationService reservationService;
 
-    public ReservationController(ReservationDao reservationDao, ReservationTimeDao reservationTimeDao) {
-        this.reservationDao = reservationDao;
-        this.reservationTimeDao = reservationTimeDao;
+    public ReservationController(ReservationService reservationService) {
+        this.reservationService = reservationService;
     }
 
     @GetMapping("/reservations")
     public ResponseEntity<List<ReservationDetailDto>> getReservations() {
-        List<ReservationDetailDto> responseData = reservationDao.selectAll()
+        List<ReservationDetailDto> responseData = reservationService.readAll()
                 .stream()
                 .map(ReservationDetailDto::from)
                 .toList();
@@ -33,21 +29,15 @@ public class ReservationController {
 
     @PostMapping("/reservations")
     public ResponseEntity<ReservationDetailDto> createReservation(@RequestBody ReservationSaveDto dto) {
-        ReservationTime reservationTime = readReservationTime(dto.timeId());
-        Reservation savedReservation = reservationDao.insert(new Reservation(dto.name(), dto.date(), reservationTime));
+        Reservation savedReservation = reservationService.reserve(dto.name(), dto.date(), dto.timeId());
         ReservationDetailDto responseData = ReservationDetailDto.from(savedReservation);
         return ResponseEntity.ok(responseData);
     }
 
     @DeleteMapping("/reservations/{id}")
     public ResponseEntity<Void> deleteReservation(@PathVariable Long id) {
-        reservationDao.delete(id);
+        reservationService.cancel(id);
         return ResponseEntity.ok().build();
-    }
-
-    private ReservationTime readReservationTime(Long id) {
-        return reservationTimeDao.select(id)
-                .orElseThrow(IllegalArgumentException::new);
     }
 
 }
